refactor(hyperdrive.php): improve security by limiting JS entry points
during an earlier beta version code was introduced to define a constant which could be used to
output the version of hyperdrive into the codebase. however, this presents a potential xss attack
vector if an attacker was able to gain control of the defined constant and use it to close the doc
block and output some nefarious code prior to fetch injection. for examples see
https://vip.wordpress.com/2014/06/20/the-importance-of-escaping-all-the-things/